Reorganize code files into folders
